+++ medic-src/src/cli/mod.rs
@@ -1,3 +1,4 @@
 pub struct Flags {
+  pub auto_apply_remedy: bool,
   pub interactive: bool,
 }