+++ medic-update/src/lib.rs
@@ -7,11 +7,11 @@ use medic_src::config::Manifest;
 use medic_src::runnable::run;
 use medic_src::AppResult;
 
-pub fn run_steps(manifest: Manifest, progress: &mut retrogress::ProgressBar, flags: Flags) -> AppResult<()> {
+pub fn run_steps(manifest: Manifest, progress: &mut retrogress::ProgressBar, mut flags: Flags) -> AppResult<()> {
   match manifest.update {
     Some(test) => {
       for step in test.steps {
-        run(step, progress, &flags)?;
+        run(step, progress, &mut flags)?;
       }
       AppResult::Ok(())
     }